refactor: restructure `static_file` module (#663)
